100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@
 int _atoi(char *s)
 {
     int sign = 1;
-    int num = 0;
+    unsigned int num = 0;
 
     while (*s)
     {
@@ -28,6 +28,5 @@ int _atoi(char *s)
         s++;
     }
 
-    return (sign * num);
+    return (int)(sign * num);
 }
-
